requirements.txt
discord.py>=2.3.2
aiofiles>=23.2.0
aiohttp>=3.8.0